Fix wiggly vehicle motion: store positions as Float64
Vehicle positions were stored in the SAB as Float32 absolute Web Mercator
world metres. At Bangalore magnitudes (~8.6e6 in X) Float32 only resolves
~1 m in X / 0.125 m in Y, quantizing positions into a staircase that read
as a ~0.4 m lateral wiggle when zoomed to street level.

Widen posX/posY (and the render-side interpolation snapshot arrays) to
Float64, dropping the position rounding error below a micron. Headings
stay Float32 (sub-milliradian over +/-pi). Adds a round-trip precision
test asserting a Bangalore-magnitude position survives to <0.001 m.

+++ src/render/interpolation.ts
@@ -6,11 +6,13 @@ export interface InterpSnapshot {
   // Positions/headings captured at the previous and current observed sim ticks.
   // Rendering lerps prev → cur over the wall-clock interval between them, which
   // smooths motion regardless of (and lagging by one tick behind) the sim.
-  prevX: Float32Array;
-  prevY: Float32Array;
+  // posX/posY are Float64 to match the SAB: absolute Web Mercator world metres
+  // need more than Float32's ~1 m resolution at Bangalore magnitudes.
+  prevX: Float64Array;
+  prevY: Float64Array;
   prevHeading: Float32Array;
-  curX: Float32Array;
-  curY: Float32Array;
+  curX: Float64Array;
+  curY: Float64Array;
   curHeading: Float32Array;
   tickNumber: number;
   capturedAtMs: number;
@@ -19,11 +21,11 @@ export interface InterpSnapshot {
 
 export function createSnapshot(): InterpSnapshot {
   return {
-    prevX: new Float32Array(MAX_VEHICLES),
-    prevY: new Float32Array(MAX_VEHICLES),
+    prevX: new Float64Array(MAX_VEHICLES),
+    prevY: new Float64Array(MAX_VEHICLES),
     prevHeading: new Float32Array(MAX_VEHICLES),
-    curX: new Float32Array(MAX_VEHICLES),
-    curY: new Float32Array(MAX_VEHICLES),
+    curX: new Float64Array(MAX_VEHICLES),
+    curY: new Float64Array(MAX_VEHICLES),
     curHeading: new Float32Array(MAX_VEHICLES),
     tickNumber: -1,
     capturedAtMs: 0,
